# order-service/src/main/resources/swagger/properties-swagger.yaml
openapi: 3.0.1

info:
  title: Order Service
  description: Order Service
  version: 1.0.0

servers:
  - url: 'https://localhost:8083'

tags:
  - name: Order
  - name: Customer

components:
  securitySchemes:
    KeycloakAuth:
      $ref: './commons.yaml#/components/securitySchemes/KeycloakAuth'

paths:
  /orders:
    get:
      tags:
        - Order
      summary: Return list of order
      operationId: getOrders
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: size
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: './commons.yaml#/components/responses/OrderResponseList'
        "400":
          $ref: './commons.yaml#/components/responses/MessageResponse'
        "500":
          $ref: './commons.yaml#/components/responses/MessageResponse'

    post:
      tags:
        - Order
      summary: Create an order
      operationId: createOrder
      requestBody:
        $ref: './commons.yaml#/components/requestBodies/OrderRequestOnCreate'
      responses:
        "200":
          $ref: './commons.yaml#/components/responses/OrderResponse'
        "400":
          $ref: './commons.yaml#/components/responses/MessageResponse'
        "500":
          $ref: './commons.yaml#/components/responses/MessageResponse'

  /orders/{uuidOrder}:
    get:
      tags:
        - Order
      summary: Return one order
      operationId: getOrder
      parameters:
        - name: uuidOrder
          in: path
          required: true
          schema:
            $ref: './commons.yaml#/components/schemas/UUID'
      responses:
        "200":
          $ref: './commons.yaml#/components/responses/OrderResponse'
        "400":
          $ref: './commons.yaml#/components/responses/MessageResponse'
        "500":
          $ref: './commons.yaml#/components/responses/MessageResponse'

  /orders/summary:
    get:
      tags:
        - Order
      summary: Return list of summary expense for each customer
      operationId: getSummaryForEachCustomer
      parameters:
        - name: data-inizio
          in: query
          required: false
          schema:
            $ref: './commons.yaml#/components/schemas/Date'
        - name: data-fine
          in: query
          required: false
          schema:
            $ref: './commons.yaml#/components/schemas/Date'
      responses:
        "200":
          $ref: './commons.yaml#/components/responses/PriceSummaryResponseList'
        "400":
          $ref: './commons.yaml#/components/responses/MessageResponse'
        "500":
          $ref: './commons.yaml#/components/responses/MessageResponse'

  /customers/{userName}/orders:
    get:
      tags:
        - Customer
      summary: Return list of order for the customer
      operationId: getCustomerOrders
      parameters:
        - name: userName
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: size
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: './commons.yaml#/components/responses/OrderResponseList'
        "400":
          $ref: './commons.yaml#/components/responses/MessageResponse'
        "500":
          $ref: './commons.yaml#/components/responses/MessageResponse'

  /customers/{userName}/orders/summary:
    get:
      tags:
        - Customer
      summary: Return summary expense for a customer
      operationId: getCustomerSummary
      parameters:
        - name: userName
          in: path
          required: true
          schema:
            type: string
        - name: data-inizio
          in: query
          required: false
          schema:
            $ref: './commons.yaml#/components/schemas/Date'
        - name: data-fine
          in: query
          required: false
          schema:
            $ref: './commons.yaml#/components/schemas/Date'
      responses:
        "200":
          $ref: './commons.yaml#/components/responses/PriceSummaryResponse'
        "400":
          $ref: './commons.yaml#/components/responses/MessageResponse'
        "500":
          $ref: './commons.yaml#/components/responses/MessageResponse'